Fix url_launcher not firing result (iOS)
If you launch a URL on iOS and hit the Done button before the SafariVC initial load completes, the FlutterResult never fires since the delegate method safariViewController:didCompleteInitialLoad: is never called.

This adds a path to safariViewControllerDidFinish: to fire the FlutterResult in case of dismissal before initial load. It's careful to make sure the result can only ever fire once.